[refactor] inline_choice() を _common.py に移動
refs #51
